% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_latest_evnt.R
\name{add_latest_evnt}
\alias{add_latest_evnt}
\title{Add the latest event to a database}
\usage{
add_latest_evnt(
  latest_evnt = NULL,
  ssid = NULL,
  sheet = NULL,
  new_sheet = F,
  overwrite = F,
  ...
)
}
\arguments{
\item{latest_evnt}{A data.frame or data.table object with latest events.}

\item{ssid}{ID of a google spreadsheet.}

\item{sheet}{The name of the sheet in the google spreadsheet to write.}

\item{new_sheet}{Logical. Should a new sheet be created? by default, the
latest events will be appended on the existing sheet.}

\item{overwrite}{logical. Do you want to overwrite data on the current sheet?
default is FALSE, to append data to an existing sheet.}

\item{...}{Additional params passed to the spreadsheet's endpoint
\code{\link[googlesheets4:gs4_create]{gs4_create()}}}
}
\value{
The input ssid, as an instance of
\code{\link[googlesheets4:as_sheets_id]{as_sheets_id()}}
}
\description{
Stores the latest event to a google spreadsheet database.
}
