% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_for_response}
\alias{check_for_response}
\title{Review an API response}
\usage{
check_for_response(res = NULL)
}
\arguments{
\item{res}{A DHIS2 response object.}
}
\description{
Check that API returned a \code{JSON} element without any errrors.
}
