% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_latest_evnt.R
\name{find_latest}
\alias{find_latest}
\alias{find_latest_by_name_and_id}
\alias{find_latest_by_name}
\alias{find_latest_by_phone_number}
\title{Find latest}
\usage{
find_latest(events, from = NULL, ...)

find_latest_by_name_and_id(events)

find_latest_by_name(events)

find_latest_by_phone_number(events)
}
\arguments{
\item{events}{A data frame with events to find.}

\item{from}{A google spreadsheet ID, where to search the events from.}

\item{...}{Additional params passed to the \link[googlesheets4]{read_sheet}.}
}
\value{
A data.table object with the latest unique events for follow up.
}
\description{
Find the latest A360 event for follow up from a list of service provision
reports stored in a google spreadsheet.
}
\details{
The latest event is determined using the \verb{girl ID}, \verb{name of girl},
and \verb{phone number} for girls with a single report and \verb{date of service provision} for girls with multiple reports.

In cases where there are multiple reports for the same girl, but the newest
is missing the phone number, the newest record is updated with any available
phone number in the reports:
\itemize{
\item record 1 phone number : NA
\item record 2 phone number : XXX XXX XXXX
\item record 3 phone number : NA
}

the newest record will be assigned the phone number : XXX XXX XXXX

this is similar to cases where there different phone numbers:
\itemize{
\item record 1 phone number : NA
\item record 2 phone number : XXX XXX XXXX
\item record 3 phone number : YYY XXX YYYY
}

the newest record will be assigned both available phone numbers : XXX XXX
XXXX | YYY XXX YYYY

the phone number is validated; only records with a phone number equal to or
more than \verb{10 digits} are returned.
}
\section{Functions}{
\itemize{
\item \code{find_latest}: Find the latest event for follow up

\item \code{find_latest_by_name_and_id}: Find the latest event by girl id and name

\item \code{find_latest_by_name}: Find latest the event by name

\item \code{find_latest_by_phone_number}: Find the latest event by phone number
}}

