% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_latest_evnt.R
\name{find_latest_by_name}
\alias{find_latest_by_name}
\title{Find latest evnt by name}
\usage{
find_latest_by_name(evnt)
}
\arguments{
\item{evnt}{the duplicated evnt, A data.frame object.}
}
\value{
a list with latest unique evnt
}
\description{
Identify the latest unique evnt from evnt with duplicated names.
}
