% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_latest_evnt.R
\name{find_latest_by_name_and_id}
\alias{find_latest_by_name_and_id}
\title{Find latest evnt by girl id and name}
\usage{
find_latest_by_name_and_id(evnt)
}
\arguments{
\item{evnt}{A data.frame or data.table object with the evnt to follow up.}
}
\value{
the latest unique evnt if present.
}
\description{
A systematic search of evnt to identify the latest unique evnt for follow
up using the girl ID and name.
}
