% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_find_latest.R
\name{find_latest_by_phone_number}
\alias{find_latest_by_phone_number}
\title{Find latest events by phone number}
\usage{
find_latest_by_phone_number(evnt)
}
\arguments{
\item{evnt}{A data.frame or data.table object containing the latest events.}
}
\value{
the latest events by phone number.
}
\description{
Find latest events by phone number
}
