% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_latest.R
\name{find_latest_event}
\alias{find_latest_event}
\title{Identify latest events}
\usage{
find_latest_event(events)
}
\arguments{
\item{events}{A dataframe.}
}
\value{
A dataframe of the latest unique record
}
\description{
Identify latest events
}
