% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_latest_evnt.R
\name{find_latest_evnt}
\alias{find_latest_evnt}
\title{Find latest evnt for follow up}
\usage{
find_latest_evnt(evnt, from = NULL, ...)
}
\arguments{
\item{evnt}{A data frame with the evnt to search from.}

\item{from}{Specify a google spreadsheet ID, where to search the evnt.}

\item{...}{Additional params passed to the \link[googlesheets4]{read_sheet}.}
}
\value{
the latest unique record, a data.table.
}
\description{
Find the latest evnt for follow-up through a systematic search of evnt
using the girl ID, name and phone number.
}
