% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_find_latest.R
\name{find_latest_unique_event}
\alias{find_latest_unique_event}
\title{Identify latest unique events for follow up}
\usage{
find_latest_unique_event(evts)
}
\arguments{
\item{evts}{the duplicated events, A data.frame object.}
}
\value{
a list with latest unique event
}
\description{
Identify latest unique events from events with duplicated names
}
