% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_girl_uid}
\alias{generate_girl_uid}
\title{Generate a unique girl ID}
\usage{
generate_girl_uid(code_size = 14)
}
\value{
A character string of size 15
}
\description{
Generate a unique identifier for girls enrolled in the follow up program.
}
\details{
\code{generate_uid} generates a random id of size 15 chars.
}
