% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_tei.R
\name{generate_payload}
\alias{generate_payload}
\title{Generate TEI payload}
\usage{
generate_payload(evnt)
}
\arguments{
\item{evnt}{A list of events from follow up database}
}
\value{
A list, TEI payload
}
\description{
Create a trackedEntityInstance from a line list
}
