% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_latest_events.R
\name{generate_random_code}
\alias{generate_random_code}
\title{Generate a random spreadsheet name}
\usage{
generate_random_code()
}
\description{
Generate a random spreadsheet name to uniquely identify the automatically
generated sheets. The sheet names starts with a360connect-db, and are then
followed with a random code.
}
