% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_teis.R
\name{generate_tei_payload}
\alias{generate_tei_payload}
\title{Generate TEI payload}
\usage{
generate_tei_payload(evnt)
}
\arguments{
\item{evnt}{A list of events to generate a payload.}
}
\value{
A list, TEI payload
}
\description{
Create a tracked entity instance payload from a line list
}
