% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_teis.R
\name{generate_teis}
\alias{generate_teis}
\title{Generate TEIs from A360 Events Program}
\usage{
generate_teis(baseurl = NULL, program_id = NULL, warn_duplicates = T)
}
\arguments{
\item{baseurl}{the server url.}

\item{program_id}{the ID of A360 Attendance & Service Log program.}

\item{warn_duplicates}{logical. Whether to highlight duplicates or not. Default is TRUE.}
}
\value{
An S3 object of the generated TEIs with possible duplicates if found.
}
\description{
Generate Tracked Entity Instances (TEIs) from A360 Attendance & Service Log,
an event program.
}
\details{
\code{generate_teis} pulls A360 events from the Attendance & Service Log,
checks for the latest unique events, and converts them into tracked entity
instances for follow up.
}
\examples{
generate_teis(baseurl = "", program_id = "x")
}
