% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_evnt.R
\name{get_evnt}
\alias{get_evnt}
\title{Extract evnt from a program}
\usage{
get_evnt(
  baseurl = NULL,
  program_id = NULL,
  startDate = NULL,
  endDate = NULL,
  user = NULL,
  pass = NULL,
  paging = TRUE,
  pageSize = 50,
  ...
)
}
\arguments{
\item{baseurl}{the server url.}

\item{program_id}{the program ID.}

\item{startDate}{the date in `YYYY-MM-DD` format.}

\item{endDate}{the date in `YYYY-MM-DD` format.}

\item{user}{string. the user account used in the pass key.}

\item{pass}{string. password of the user account, used in the pass key.}

\item{paging}{logical. Whether to page the results or not. By default, the
return evnt are paged.}

\item{pageSize}{numeric. indicating the size of the page. default is `50`.}
}
\value{
A S3 object containing a content, endpoint and the parsed response.
}
\description{
Extract evnt from a DHIS2 program.
}
