% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_teis.R
\name{get_tei}
\alias{get_tei}
\title{Get Tracked Entity Instance (TEI) from a DHIS2 server}
\usage{
get_tei(tei_id, baseurl = NULL, user = NULL, pass = NULL)
}
\arguments{
\item{tei_id}{The ID of a Tracked Entity Instance to pull.}

\item{baseurl}{The DHIS2 server, default is psi-clone server.}

\item{user}{DHIS2 User Account, used in the pass key.}

\item{pass}{Password of the user account, used in the pass key.}
}
\value{
A DHIS2 response
}
\description{
Get Tracked Entity Instance (TEI) from a DHIS2 server
}
