% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{has_duplicate_names}
\alias{has_duplicate_names}
\title{Check if any record has a duplicate name}
\usage{
has_duplicate_names(events)
}
\arguments{
\item{events}{A data.frame object}
}
\value{
Logical
}
\description{
Check if any record has a duplicate name
}
