% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_latest.R
\name{has_phone_number}
\alias{has_phone_number}
\title{Check if the any record has a phone number}
\usage{
has_phone_number(events)
}
\value{
Logical
}
\description{
Check if the any record has a phone number
}
