% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_teis.R
\name{import_teis}
\alias{import_teis}
\alias{create_and_update_teis}
\alias{create_teis}
\alias{update_teis}
\title{Import Tracked Entity Instances (TEIs) on a DHIS2 server}
\usage{
import_teis(teis, strategy = c("create_and_update", "create", "update"), ...)

create_and_update_teis(teis, ...)

create_teis(
  teis,
  baseurl = NULL,
  user = NULL,
  pass = NULL,
  tracked_entity_type = NULL,
  tea = NULL,
  tea_mapping = NULL,
  is_a360connect_type = TRUE,
  is_a360connect_enrollment_type = TRUE
)

update_teis(
  teis,
  baseurl = NULL,
  user = NULL,
  pass = NULL,
  tracked_entity_type = NULL,
  tea = NULL,
  tea_mapping = NULL,
  is_a360connect_type = TRUE,
  is_a360connect_enrollment_type = TRUE
)
}
\arguments{
\item{teis}{A data.frame or data.table object containing a list of tracked
entity instances to import.}

\item{strategy}{A character string specifying the import strategy. Currently
only three options are specified, \code{create_and_update}, \code{create} and
\code{update}. The default is \code{create_and_update}.}

\item{...}{Additional params passed to \code{\link[=create_teis]{create_teis()}} and \code{\link[=update_teis]{update_teis()}}.}

\item{baseurl}{The DHIS2 server, default is psi-clone server.}

\item{user}{DHIS2 User Account, used in the pass key.}

\item{pass}{Password of the user account, used in the pass key.}

\item{tracked_entity_type}{DHIS2 ID, the type of tracked entity attribute.}

\item{tea}{A vector, specifies the tracked attributes.}

\item{tea_mapping}{A data.frame object containing the tracked entity attributes and their uids.}

\item{is_a360connect_type}{Logical. Use this to bypass \code{a360connect} settings.}

\item{is_a360connect_enrollment_type}{Logical, Use this to bypass \code{a360connect} settings.}
}
\value{
An S3 object with a DHIS2 response, API endpoint and content of the
parsed response.
}
\description{
A generic function to upload Tracked Entity Instances on a DHIS2 server.
}
\details{
\code{\link[=import_teis]{import_teis()}} uploads a list of tracked entity instances to a
DHIS2 server. It provides options to specify the import strategy, point to
a DHIS2 instance using the baseurl, and pass authentication. By default,
the basic authentication is supported, and all the teis will be created or
updated where possible.

\code{\link[=create_and_update_teis]{create_and_update_teis()}} Checks for the existence of a tracked entity
instance on a DHIS2 server before sending a request to create or update the
object. By default, the tracked entity instances are created or updated on
the psi-clone server - this can be changed by supplying the baseurl.

\code{\link[=create_teis]{create_teis()}} Generates the payload directly from a data.frame with a
list of tracked entity instances, and posts them to a DHIS2 instance. By
default, the tracked entity instances are posted to the psi-clone server -
this can be changed by supplying the baseurl.

\code{\link[=update_teis]{update_teis()}} Generates the payload directly from a data.frame with a
list of existing tracked entity instances, and updates them on a DHIS2
instance. By default, it updates the tracked entity instances to the
psi-clone server - this can be changed by supplying the baseurl.

The baseurl and authentication, if not specified explicitly, will be pulled
from the \code{r} \code{environment}.
}
\section{Functions}{
\itemize{
\item \code{import_teis}: Imports Tracked Entity Instances (TEIs) on a DHIS2
server

\item \code{create_and_update_teis}: Create and Update Tracked Entity Instances (TEIs) on
a DHIS2 server

\item \code{create_teis}: Create Tracked Entity Instances (TEIs) on a DHIS2
server

\item \code{update_teis}: Update Tracked Entity Instances (TEIs) on a DHIS2
server
}}

\examples{
\dontrun{
import_teis(teis)
server_url <- "https://dhis-instance.org/"
user <- "admin"
pass <- "district"
# creates and updates the teis
import_teis(teis, baseurl = server_url, user = user, pass = pass)
# creates the teis
import_teis(teis, strategy = "create", baseurl = server_url, user = user, pass = pass)
# updates the teis
import_teis(teis, strategy = "update", baseurl = server_url, user = user, pass = pass)
}
\dontrun{
create_and_update_teis(teis)
server_url <- "https://dhis-instance.org/"
user <- "admin"
pass <- "district"
create_and_update_teis(teis, baseurl = server_url, user = user, pass = pass)
}
\dontrun{
create_teis(teis)
server_url <- "https://dhis-instance.org/"
user <- "admin"
pass <- "district"
create_teis(teis, baseurl = server_url, user = user, pass = pass)
}
\dontrun{
update_teis(teis)
server_url <- "https://dhis-instance.org/"
user <- "admin"
pass <- "district"
update_teis(teis, baseurl = server_url, user = user, pass = pass)
}
}
