% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_api_response}
\alias{parse_api_response}
\title{Parse an API response}
\usage{
parse_api_response(res, url, simplify_vector = F, name = NULL)
}
\arguments{
\item{res}{the API response}

\item{url}{the endpoint url}

\item{simplify_vector}{passed to \link{jsonlite::fromJSON}}

\item{name}{class name of the S3 object}
}
\value{
S3 object
}
\description{
Parse an API response
}
