% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_teis.R
\name{pull_events}
\alias{pull_events}
\title{Pull Events}
\usage{
pull_events(
  baseurl = NULL,
  program_id = NULL,
  startDate = NULL,
  endDate = NULL
)
}
\arguments{
\item{baseurl}{the server url.}

\item{program_id}{the ID of A360 Attendance & Service Log program.}
}
\description{
Pull Events
}
