% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{review_search_result}
\alias{review_search_result}
\title{Review search results}
\usage{
review_search_result(evts)
}
\arguments{
\item{evts}{A list of data.table, the events return from search}
}
\value{
reviewed result, A data.table if
}
\description{
Combine results and return unique events
}
