% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_teis.R
\docType{data}
\name{selected_tea}
\alias{selected_tea}
\title{Selected Tracked Entity Attributes}
\format{
\describe{
\item{Name of girl}{name of a tracked entity attribute, captures the name of girl}
\item{Girl ID}{name of a tracked entity attribute, captures the A360 girl id}
\item{Phone Number}{name of a tracked entity attribute, captures the Phone Number}
\item{Method taken up}{name of a tracked entity attribute, records the method taken up}
\item{Date of Service Provision}{name of a tracked entity attribute, records the date of service provision}
\item{Follow-up scheduled (date)}{name of a tracked entity attribute, captures the scheduled date of the follow up call}
\item{KEY}{refers to the Unique Identifier for the TEIs, captured in \verb{CORE - Unique ID (UIC) }}
}
}
\usage{
selected_tea
}
\description{
A vector containing a names of tracked entity attributes implemented in \code{A360}
\verb{provider Call Log Program}
}
\keyword{datasets}
