% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_teis.R
\docType{data}
\name{tea_map}
\alias{tea_map}
\title{Tracked Entity Attributes data map}
\format{
\describe{
\item{id}{id, Unique ID of the tracked entity attribute}
\item{name}{name,name of the tracked entity attribute}
}
}
\source{
\url{https://staging.psi-mis.org/}
}
\usage{
tea_map
}
\description{
A dataset containing the ids and name tracked entity attributes used in \code{A360}
\verb{Provider Call Log program}
}
\keyword{datasets}
