% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_teis.R
\name{transform_event_datavalues}
\alias{transform_event_datavalues}
\title{Transform event data values}
\usage{
transform_event_datavalues(events)
}
\arguments{
\item{events}{A data.frame object.}
}
\description{
Transform event data values
}
