% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_evnt.R
\name{transform_events}
\alias{transform_events}
\title{Transform event}
\usage{
transform_events(events)
}
\arguments{
\item{events}{A data.frame object with events to transform.}
}
\description{
Transform event data values from long to wide, and selects the required
fields for follow-up. In case the events are missing data values, only the
required fields are selected.
}
