% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_latest_evnt.R
\name{add_latest_evnt}
\alias{add_latest_evnt}
\title{Add the latest evnt to a database}
\usage{
add_latest_evnt(
  latest_evnt = NULL,
  ssid = NULL,
  sheet = NULL,
  new_sheet = F,
  overwrite = F,
  ...
)
}
\arguments{
\item{latest_evnt}{A data.frame object.}

\item{ssid}{the ID of a google spreadsheet.}

\item{sheet}{the name of the sheet}

\item{new_sheet}{logical. Create a new sheet? default is FALSE, to append
data to an existing spreadsheet.}

\item{overwrite}{logical. Do you want to overwrite data on the current sheet?
default is FALSE, to append data to an existing sheet.}

\item{...}{Additional params passed to the spreadsheet's endpoint
\ref{googlespreadsheet4::gs4_create}.}
}
\value{

}
\description{
Stores the latest evnt to a database. Currently, it stores the records in a
google spreadsheet.
}
