% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_find_latest.R
\name{find_latest_by_name}
\alias{find_latest_by_name}
\title{Find latest events by name}
\usage{
find_latest_by_name(evnt)
}
\arguments{
\item{evnt}{the duplicated events, A data.frame object.}
}
\value{
a list with latest unique event
}
\description{
Identify the latest unique events from events with duplicated names.
}
