% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_find_latest.R
\name{find_latest_by_name_and_id}
\alias{find_latest_by_name_and_id}
\title{Find latest events by girl id and name}
\usage{
find_latest_by_name_and_id(evnt)
}
\arguments{
\item{evnt}{A data.frame or data.table object with the events to follow up.}
}
\value{
the latest unique events if present.
}
\description{
A systematic search of events to identify the latest unique events for follow
up using the girl ID and name.
}
