% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_latest_evnt.R
\name{find_latest_by_phone_number}
\alias{find_latest_by_phone_number}
\title{Find latest evnt by phone number}
\usage{
find_latest_by_phone_number(evnt)
}
\arguments{
\item{evnt}{A data.frame or data.table object containing the latest evnt.}
}
\value{
the latest evnt by phone number.
}
\description{
Find latest evnt by phone number
}
