% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_find_latest.R
\name{find_latest_event}
\alias{find_latest_event}
\title{Find latest events}
\usage{
find_latest_event(events)
}
\arguments{
\item{events}{A dataframe.}
}
\value{
A data frame of the latest unique record
}
\description{
Identify the latest events for follow-up.
}
\details{
Reviews an entire list of events and identify the latets unique events.
}
