% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_latest_evnt.R
\name{generate_girl_uid}
\alias{generate_girl_uid}
\title{Generate a unique girl ID}
\usage{
generate_girl_uid(code_size = 14)
}
\arguments{
\item{code_size}{Integer, size of the uid. default 14.}
}
\value{
A character string, the unique girl ID
}
\description{
Generate a unique girl ID to identify the events added to a database.
}
\details{
Randomly generates a unique code, with three parts each separated with an
hyphen. XXXXXX-XYXY-00000.
}
