% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_teis.R
\name{generate_tei_payload}
\alias{generate_tei_payload}
\title{Generate TEI payload}
\usage{
generate_tei_payload(
  evnt,
  tracked_entity_type = NULL,
  tea = NULL,
  tea_mapping = NULL,
  is_a360connect_type = TRUE,
  is_a360connect_enrollment_type = TRUE
)
}
\arguments{
\item{evnt}{A data.frame of events to generate a payload.}

\item{tracked_entity_type}{DHIS2 ID, the type of tracked entity attribute.}

\item{tea}{A vector, specifies the tracked attributes.}

\item{tea_mapping}{A data.frame object containing the tracked entity attributes and their uids.}

\item{is_a360connect_type}{Logical. Use this to bypass \code{a360connect} settings.}

\item{is_a360connect_enrollment_type}{Logical, Use this to bypass \code{a360connect} settings.}
}
\value{
A list, TEI payload
}
\description{
Create a tracked entity instance payload from a line list
}
