% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_events.R
\name{modify_events_endpoint}
\alias{modify_events_endpoint}
\title{Modify events Endpoint}
\usage{
modify_events_endpoint(
  baseurl = NULL,
  program_id = NULL,
  startDate = NULL,
  endDate = NULL,
  paging = TRUE,
  pageSize = 50,
  pageNumber = 1
)
}
\arguments{
\item{baseurl}{the server url.}

\item{program_id}{the program ID.}

\item{startDate}{the date in \code{YYYY-MM-DD} format.}

\item{endDate}{the date in \code{YYYY-MM-DD} format.}

\item{paging}{logical. Whether to page the results or not. By default, the
return events are paged.}

\item{pageSize}{numeric. indicating the size of the page. default is \code{50}.}

\item{pageNumber}{Numeric. Specifies the page number to be returned. default is page \code{1}.}
}
\value{
url
}
\description{
Code to modify events API Endpoint
}
