% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_pull.R
\name{transform_events}
\alias{transform_events}
\title{Transform event data values}
\usage{
transform_events(events)
}
\arguments{
\item{events}{A data.frame object.}
}
\description{
Transform event data values
}
