% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_events.R
\name{transform_events}
\alias{transform_events}
\title{Transform events}
\usage{
transform_events(events)
}
\arguments{
\item{events}{A data.frame object with events to transform.}
}
\description{
Transform events data values from long to wide, and selects the required
fields for follow-up.
}
